repair vercel
